Optimize to calculate TLightShader.Code on demand.
When shaders are not used, then ApplyInternalShaders is not called, and then the shader source code for most lights will not have to be calculated. This is a huge optimization (unfortunately, lame, since only for non-shader pipeline), as calculating light shader code takes a lot of time.

Note that code for lights used by shadow maps will still be calculated, but in this case it's not that often, and shapes with shadow maps probably will want shaders anyway.

git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/kambi_vrml_game_engine@8427 afe8dcc5-9094-451b-8773-5f495042a737