Remove TVRMLInterfaceDeclaraion.UniformInvalid.
It was unclean to store such info inside VRML graph. For example, it causes bugs: validness needs to be reconsidered e.g. when user changes shader URL.

Instead, introduce UniformsTextures instead of UniformNodes, and actually remove the textures inside the shader. This is clean, as UniformsTextures if a field of shader program.

git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/kambi_vrml_game_engine@8961 afe8dcc5-9094-451b-8773-5f495042a737