Removed examples/images/dds_remove_small_mipmaps.
This was a tool to remove from DDS mipmaps the levels with one of the sizes < 4. It workarounds bug in older version of GIMP-DDS, it generated invalid images for smaller mipmaps. The bug is fixed in GIMP-DDS >= 2.0.9. See http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=564111 .

Since the fixed version of GIMP-DDS is available since a long time (and the tool didn't really serve any other purpose), it's removed now.


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/kambi_vrml_game_engine@9813 afe8dcc5-9094-451b-8773-5f495042a737