Rename TCamera.NavigationInput to simpler TCamera.Input, along with related type name.
Rename ciKeyboard (previously niKeyboard) to ciNormal. That's because all Input_Xxx shortcuts describe both key and mouse inputs, and they affect both key/mouse-button/mouse-wheel holding (when Idle does something continous when key is down) or key/mouse-button/mouse-wheel clicking (when EventDown does something once when input is pressed).
So, the new name "ciNormal" is quite poor and non-descriptive, but that's because it really toggles various input devices and reactions.

git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle_game_engine@11269 afe8dcc5-9094-451b-8773-5f495042a737