Remove very dangerous frRootNode feature.
It was useless --- all tests showed that RootNode graph isn't the main memory eater on models now. Also, it was only for models absolutely static (never ProcessEvents=false).

The argument that finally proved that it should be removed: frRootNode needs to do BeforeNodesFree, as Transform nodes must be removed from TransformInstancesList, otherwise freeing scene later will break. Which means that it must be followed by ChangedAll, which means it's slow and more useless for optimizing already loaded models.


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle_game_engine@11397 afe8dcc5-9094-451b-8773-5f495042a737