Fix TCastleSceneManager.MouseRayHit storage: we have to observe all 3D items that we're picked by last ray, in case any of them is freed -> we have to free MouseRayHit list.
This fixes the case when we free (by RemoveAndFree) a 3D item from a hierarchy, and it happens to be the same 3D item that was picked by the last operation, and it's not the first item on MouseRayHit list. We previously had special code (MouseRayHit3D) for the first 3D item on MouseRayHit list, now we just observe them all in the same way.


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle_game_engine@11436 afe8dcc5-9094-451b-8773-5f495042a737