Every T3DOrient automatically has now a Camera inside, and stores it's position/direction/up vector inside.
This greatly simplifies whole TPlayer <-> Camera synchronization, which now just happens inside T3DOrient, and is just magically in both direction (because the data is really stored only in TWalkCamera instance, it's not replicated).

This removes the need for synchronization in TPlayer.Translate (T3DOrient.Translate already does the job) or TCastleSceneManager.CameraMoveAllowed. It also means that assigning to Player.Position works, which restores teleports working in castle1.


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle_game_engine@11890 afe8dcc5-9094-451b-8773-5f495042a737