CastleGLUtils, CastleGLImages, CastleGLShaders, CastleControls, CastleGLBitmapFonts compile for OpenGL ES :)
Note that TGLImage and TGLBitmapFont don't render anything at all for now.

Also, preparations for OpenGL ES in other units: cube map consts (GL_TEXTURE_CUBE_MAP_ARB and so on) lose the ARB suffix. For GLES20, only the versions without suffix are available. (And for desktop OpenGL, versions with and without suffix are equal.)


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle_game_engine@13179 afe8dcc5-9094-451b-8773-5f495042a737