Implemented dmMultiply blending mode
dmMultiply blending mode for TRGBAlphaImage. This is simple implementation of 'multiply' filter like in GIMP which I needed to add some diversity to my game UI leaving it scalable. It seems that it should work with any destination/source image type, but requires some testing to see if everything is working as expected. I've added the dmMultiply everywhere where it was applicable in TRGBAlphaImage.drawFromCore (and if I was sure what I am doing :)), but I've tested it only for TRGBAlphaImage to TRGBAlphaImage.