Add -w to ignore linker warnings
This seems the only way to get rid of XCode (>= 8.3) linker errors when
compiling iOS project. The error is

  Warning:pointer not aligned at address...

and it is caused when compiling x86 code for iPhone Simulator.
The error is inside FPC RTL, FPC developers consider this warning
unnecessary, aligning the relevant structured would be wasteful:
https://bugs.freepascal.org/view.php?id=31696

See more:
https://forum.lazarus.freepascal.org/index.php?topic=36978.0
https://stackoverflow.com/questions/41229076/project-settings-recommends-compiler-warning-suspicious-moves