CastleVectors, CastleClassUtils no longer depend on CastleDownload
This simplifies unit dependencies, previously we had a bit of "tangled
knot", causing the compilation of CastleVectors (which should be a unit
"at the bottom") to recompile many other units in net/, including
CastleDownload.

Necessary changes:

- TTextReader/Writer, CreateReadFileStream, StreamSaveToFile are
  now in CastleDownload.
- StdInReader is removed (you can easily initialize such thing
  yourself, this is just "StdInReader := TTextReader.Create(StdInStream, false)")
- FileToString and StringToFile moves to CastleFilesUtils

This way CastleVectors, CastleClassUtils, CastleLog and others no
longer depend on the "net" stuff.