Different strategy for warning "opening files before Application.OnInitialized"
The FileAccessSafe is by default true. It is only changes to false in CastleWindow initialization, and then changes back to true in CastleWindow before Application.OnInitialize.

This way this warning is not troublesome at all for programs not using CastleWindow, e.g. using CastleControl, or command-line (like tovrmlx3d) or using custom OpenGL context creation.