Remake TSoundBuffer to be a class, that may have ALBuffer = 0 (not initialized for OpenAL)
This allows a TSoundBuffer instance to "survive" when Android application is paused + resumed, which causes OpenAL close + reopen. The OpenAL buffers are destroyed by this, and need to be reloaded. The new TSoundBuffer has all the information to reload itself in ALContextOpen.

Also code is cleaner this way:

- TSoundBuffer may expose Duration property is a trivial way now.

- TSoundBuffer can be used instead of TSoundBufferCache.

Also various docs improved in CastleSoundEngine, e.g. TSoundAllocator and ALContextOpen/Close.