Detection of CodeTyphon FPC with generics compatibility.
CodeTyphon has its own FPC version with small modifications. One thing is its own Generics.Collection in a seperate package. That is why the build tool could not find the units. With a detection if CodeTyphon is installed, now, in this case, it will load the compatibility units like in FPC versions below 3.1.1.