Fix drawing_modes_test.lpr image visibility
The previous version was accidentally hiding the images, because it
was creating TCastleWindow instead of TCastleWindowCustom. And
TCastleWindow has a default Window.SceneManager that by default is
just filled with a black background. This Window.SceneManager was
unused... but it was displayed on top of images (added by
Window.Controls.InsertBack).

This used to work differently in old (6.0 I think) engine version that
had a complicated idea of RenderStyle causing SceneManager to be
always underneath the rest. But RenderStyle is not used anymore, it
was more confusing than helpful usually.

To keep the window background defined, we now use
TCastleRectangleControl with black color. This is simple and reliable.