"castle-engine auto-generate-textures" uses now MD5 hashes
The hashes are used to detect if a texture needs updating (downscaling
or compression detection), for
https://castle-engine.io/creating_data_auto_generated_textures.php .
Using content hashes is much more reliable than using file
modification times, that are easily not suitable for this
decision after git/svn update.

The hashes are stored in new castle_engine_auto_generated.xml file.