Calculate more TUIControl position/rects using floats, not ints
This makes calculation better in case of UI scaling, that previously
was forced to use many Round() in mid-calculation, thus the end result
could "jitter". Instead, we should use just 1 Round() (at the end),
or even none (if we will use sub-pixel rendering for UI one day -- not
now), and in the middle calculate everything using floats.