Allow to customize TCastleImageControl.DrawableImage, OwnsDrawableImage
Includes also code simplifications to how TCastleImageControl deals
with FDrawableImage (previous FGLImage). It is now automatically
created in the constructor of TCastleImageControl, and assumed to be
non-nil throughout the whole instance life. (The previous approach,
freeing it in GLContextClose, was a historical artifact.)

Various TCastleImageControl are now directly synchronized with FDrawableImage.