TCastleWindow/ControlCustom -> TCastleWindow/ControlBase
This should be more encouraging name to use this class. Most
non-trivial applications should use TCastleWindow/ControlBase instead
of TCastleWindow/Control in the future.

The suffix "Custom" was falsely suggesting that this class is not
supposed to be used (in LCL, "Custom" generally means that the class
has no published properties, and it's useful to make descendants).