Remove TCastleFont.Scale, it was unreliable (the way it was implemented)
Changing Scale was OK for TTextureFont, but not for TFontFamily or
TCustomizedFont. The Scale was meant as internal mechanism (to allow
to scale TTextureFont, to allow various sizes for TTextureFont),
not as something settable on any TCastleFont.

All the code should modify only TCastleFont.Size now.

Note that TCastleLabel still exposes both FontSize and FontScale.
The FontScale is internaly applied OK (without using TCastleFont.Scale).

In the future we may add *proper* TCastleFont.Scale, that just works everywhere.