Do not use InitialNode to "match" the texture cache
Fixes a bug submitted by Eugene Loza about wrong texture scale in some
circumstances in "The Unholy Society" fight.
It was actually the wrong texture displayed (the scale is right, but
for a different image).

Reason, in short:

The original X3D node (that caused creation of a texture cache slot)
was destroyed, but a new texture X3D node (with different URL, but the
same reference) was created. Our condition
"""TextureCached.InitialNode = TextureNode""" meant that we used for
the new node the OpenGL resource associated with the old node, so we
showed a wrong texture.

History and analysis:

This was a bug I made on 12 May 2006, 13 years ago :) I hunted it down
to
https://github.com/castle-engine/castle-engine/commit/5c3557c887b81f8ddc64dcad3e1f1fc2a80e775a
, which is the first implementation of "cache" for OpenGL texture
resources.

There I assumed that X3D node reference (in
`TTextureCache.Node`) will always be "reliable", i.e. if another node
has the same reference then it's obviously the same node and can reuse
the same cache.

Which is not true, if the previous node ("original" node that caused
the creation of this cache slot) can be destroyed, and another texture
node (with different texture URL, but same reference) can be
created.

I was hunting down for a reason *why* did I even do this comparison,
and the above commit (13 years ago) explains it. This code
was refactored many times since 2006, the `UseTextureFileNames`
property was removed since that time, and I should have just removed
`InitialNode` at that point. Instead I added comparison that qualified
as "match" two nodes with equal references, without looking at their
URLs.