New approach to define list by callbacks for TSerializationProcess.ReadWrite
This is necessary to call AddBehavior (not just directly insert to list new behavior, without setting TCastleBehavior.Parent). In general this allows to put some burden on each class, instead of forcing TSerializationProcess to account for all cases. E.g.

- filtering by component state,

- or creating list on-demand,

are no longer something that is done in TSerializationProcess descendants.