Not using ndk-build anymore (with it's Makefile-based system) as it doesn't handle project with spaces
We don't actually need it at all (no need for CMake either) as we only provide precompiled libraries. The fact we used ndk-build was a historic artifact -- first it was needed always, then it was needed to get ndk-gdb working (but it doesn't anyway). We can now remove ndk-build without losing anything.