Removed serializedvectors.pas from physics_persistent_forces_components
This was making an exception how we expose vectors as published. Removed now -- the benefit over TCastleVector3Persistent is only that you don't need getter/setter boilerplate, but at the cost of now being able to define any setter for vector (and worse syntax for access). This is not good enough for CGE purposes.