Move StringMatchesRegexp to CastleTestUtils
We do not want to support it in CGE for users, as current implementation is incompatible between FPC, Delphi and even FPC 3.2.0 and 3.2.2. It's quite OK for current usage (to compare tests in TTestX3DNodes), it's not enough for general usage.

If in the future we'll want to expose this to users, we should drop FPC 3.2.0 support. Or even include own copy of RegExpr for both FPC and Delphi.