Fix (and simplify) cooperation between WholeSceneManifold and DistanceCulling
When WholeSceneManifold, then distance culling cannot be done per-shape. That's it, local change in TCastleScene.DistanceCullingCheckShape. No need for various InternalForceRendering checks, that were not necessary (when something is not in frustum, it doesn't have to be rendered) and also were broken (as LocalRender takes place before LocalRenderShadowVolumes).

This also fixes distance culling + WholeSceneManifold when used with TCastleTransformReference, added to test.