Remove TCastleTableView, deprecated for a long time, with uncomfortable API
The API should rather expose methods to add rows and columns. Like simple API of TCastleVertical/HorizontalGroup.

Ideally, we should have a grid component, that allows to add stuff in an easy way like this:

    Grid := TCastleGrid.Create(Owner);
    Grid.Rows := 5;
    Grid.Columns := 3;
    Grid.Cells[0, 0] := Label1;
    Grid.Cells[0, 2] := Label2;

  or

    Grid := TCastleGrid.Create(Owner);
    Grid.Columns := 3;
    GridRow1 := TCastleGridRow.Create(Owner);
    GridRow1.InsertFront(Label1);
    GridRow1.InsertFront(Label2);
    GridRow1.InsertFront(Label3);
    Grid.InsertFront(GridRow1);

The current TCastleTableView required to define table contents by defining a class that satisfies specific interface. This is less comfortable.